make some massive selectors for game loadout metadata
