+++ src/app/loadout-drawer/loadouts-selector.ts
@@ -1,0 +1,15 @@
+import { currentProfileSelector } from 'app/dim-api/selectors';
+import { RootState } from 'app/store/types';
+import { emptyArray } from 'app/utils/empty';
+import { createSelector } from 'reselect';
+import { convertDimApiLoadoutToLoadout } from './loadout-type-converters';
+import { Loadout } from './loadout-types';
+
+/** All loadouts relevant to the current account */
+export const loadoutsSelector = createSelector(
+  (state: RootState) => currentProfileSelector(state)?.loadouts,
+  (loadouts) =>
+    loadouts
+      ? Object.values(loadouts).map((loadout) => convertDimApiLoadoutToLoadout(loadout))
+      : emptyArray<Loadout>()
+);
