+++ src/app/loadout/ingame/InGameLoadoutDetailsSheet.m.scss
@@ -1,5 +1,5 @@
 .sheet {
-  width: fit-content;
+  width: min-content;
   margin: auto;
 }
