+++ src/app/loadout/ingame/InGameLoadoutStrip.m.scss.d.ts
@@ -1,8 +1,6 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'addLoadoutIndexNumber': string;
-  'addLoadoutTileWrapper': string;
   'blankBox': string;
   'equipNok': string;
   'equipOk': string;
